MFC r296909: Fix witness panic in the ipoib_ioctl() function when unloading the ipoib module.
The bpfdetach() function is trying to turn off promiscious mode on the
network interface it is attached to while holding a mutex. The fix
consists of ignoring any further calls to the ipoib_ioctl() function
when the network interface is going to be detached. The ipoib_ioctl()
function might sleep.

Sponsored by:	Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/9@297649 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f