Use NULL instead of 0 for pointers.
fgetln(3) will returns NULL if cannot get a line from a stream.
strsep(3) it will returns NULL if the end of the string was reached.
jemalloc(3) malloc will returns NULL if it cannot allocate memory.
fgetln(3) it will returns NULL if it cannot get a line from a stream.

MFC after:	4 weeks


git-svn-id: https://svn.freebsd.org/base/head@298033 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f