Merge the projects/release-pkg branch to head.
This allows packaging the base system with pkg(8), including
but not limited to providing the ability to provide upstream
binary update possibilities for non-tier-1 architectures.

This merge is a requirement of the 11.0-RELEASE, and as such,
thank you to everyone that has tested the project branch.

Documentation in build(7) etc. is still somewhat sparse, but
updates to those parts will follow.

Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/head@298107 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f