Restore histrory messed up in r290620
ascii.c was removed during r290494 but this introduced an issue with ASCII that
has been fixed in r290618 and lead to a rewrite of ascii.c based on none.c.

ascii.c was wrongly reintroduced in r290620 without proper svn operation which
lost the history.

Reported by:	ache, danfe


git-svn-id: https://svn.freebsd.org/base/head@298395 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f