sys: use our roundup2/rounddown2() macros when param.h is available.
rounddown2 tends to produce longer lines than the original code
and when the code has a high indentation level it was not really
advantageous to do the replacement.

This tries to strike a balance between readability using the macros
and flexibility of having the expressions, so not everything is
converted.


git-svn-id: https://svn.freebsd.org/base/head@298433 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f